requirements.txt
Flask==2.2.2
Flask-Caching==2.0.1
snowflake-connector-python==2.8.0
pandas==1.5.3
requests==2.28.1
gunicorn==20.1.0
locust